eval-annotations: such a script should never be interactive
Scripts which are intended to be used via 'repo forall -c' commands
should not be interactive as there are still the possibility to launch
those scripts concurrently with the parameter '-j' of repo.

Moreover, it seems a better idea to create a dedicated function
"push-upstream-mirror-branches" to answer the need of pushing the upstream
mirror branches created via "fetch-upstream" function.

Change-Id: I940487783221e8fa6c2b19f04ac9faeccd4442ce