CLIP OS toolkit environment: Split up setup and activation steps
This split between the setup (i.e. creation from scratch) of the CLIP
OS toolkit environment (an enhanced Python virtualenv) and the
activation process are made to avoid race condition issues. This is
explained by the fact that the activation step can be safely done in
parallel as it does not affect any file in the CLIP OS source tree
per-se and only affects parameters of the current shell session.

Another advantage of this split is that the activation step is now a
bit quicker since "setup.sh" (which was previously "setup_venv.sh") is
not called automatically and left to the user intention. The new
"activate" script (previously "source_me.sh") does its best to warn
the user when he/she has to call "setup.sh" because it is required
(either because the virtualenv has never been setup, or because the
virtualenv setup in "run/venv" is out-of-date with the new virtualenv
specifications).

Change-Id: I771356601d87ccf6b46d6360faf59194e687d3cd
Signed-off-by: Nicolas Godinho <nicolas.godinho@ssi.gouv.fr>