build: enable container based builds
This enables documentation builds using sphinx from a container based on
Debian 10. Container images are pulled from the GitLab registry.

If sphinx-build is installed on the system and the CLIPOS_USE_HOST_TOOLS
environment variable is set, then the system installed sphinx command is
used.

Else, we try using podman if it is available and we try to run
containers unprivileged if the user has UID/GID mappings in
/etc/sub{u,g}id.

Finally, we fallback on a privileged call to Docker.

Change-Id: Iec86131136e54fe69372bc36b5a2eb44b87e403d
Signed-off-by: Timothée Ravier <timothee.ravier@ssi.gouv.fr>